chore: update as per PR feedback
