+++ demo/infra/src/application/presentation/index.ts
@@ -294,8 +294,6 @@ export class PresentationStack extends NestedStack {
         }),
       ],
     });
-    // policy.attachToRole(props.authenticatedUserRole);
-    // props.authenticatedUserRole.attachInlinePolicy(policy);
 
     NagSuppressions.addResourceSuppressions(
       policy,
